requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
grpcio==1.59.2
grpcio-tools==1.59.2
protobuf==4.25.1 